# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "A comprehensive MERN stack API for managing school operations including authentication, attendance, grades, quizzes, and timetables",
    "contact": {
      "name": "API Support",
      "email": "support@schoolmanagement.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/quiz/create": {
      "post": {
        "tags": ["Quizzes"],
        "summary": "Create AI-powered quiz",
        "description": "Create a quiz with AI-generated questions using Groq API (Admin only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "subject", "chapter", "classSection"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Mathematics Chapter 1 Quiz"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Mathematics"
                  },
                  "chapter": {
                    "type": "string",
                    "example": "Trigonometry"
                  },
                  "classSection": {
                    "type": "string",
                    "example": "10A"
                  },
                  "duration": {
                    "type": "integer",
                    "description": "Duration in seconds",
                    "example": 1800
                  },
                  "numberOfQuestions": {
                    "type": "integer",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz created successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "500": {
            "description": "Failed to generate questions"
          }
        }
      }
    },
    "/quiz/available": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "Get available quizzes",
        "description": "Get all available quizzes for student's class",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of available quizzes"
          },
          "403": {
            "description": "Access denied - Students only"
          }
        }
      }
    },
    "/quiz/start/{quizId}": {
      "post": {
        "tags": ["Quizzes"],
        "summary": "Start quiz attempt",
        "description": "Start a new quiz attempt (Student only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "attemptId": {
                      "type": "string"
                    },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "duration": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Already attempted or quiz not active"
          }
        }
      }
    },
    "/quiz/submit/{attemptId}": {
      "post": {
        "tags": ["Quizzes"],
        "summary": "Submit quiz",
        "description": "Submit completed quiz (Student only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "score": {
                      "type": "number"
                    },
                    "correctAnswers": {
                      "type": "integer"
                    },
                    "totalQuestions": {
                      "type": "integer"
                    },
                    "timeTaken": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quiz/results/{attemptId}": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "Get quiz results",
        "description": "Get detailed results for a completed quiz",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "attemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz results retrieved successfully"
          },
          "400": {
            "description": "Quiz not completed yet"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}